Hello all. I'm Awumdah
I've been around the block and done a lot of interesting things. I've decided to share a bit and maybe create from within my imagination, based on my experiences. Furthermore, I'm fascinated with the new technology, and I'm on a journey to learn more. I like most of us, woke up to find ourselves caught in the net of the worldwide web. So when in Rome, right?